Fixed an issue with config validation. Now validation errors are logged. Fixed an issue with thread locallity of sqlite objects.

+++ bookworm/config/spec.py
@@ -11,7 +11,7 @@ config_spec = StringIO(
     show_file_name_as_title = boolean(default=False)
     open_with_last_position = boolean(default=True)
     auto_check_for_updates = boolean(default=True)
-    last_update_check = integer(default=0)
+    last_update_check = float(default=0)
     play_pagination_sound = boolean(default=True)
     speak_page_number = boolean(default=True)
     speak_section_title = boolean(default=True)
